+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+* Added an ability to provide list of marketplace names as a param attribute to **validate** and **upload**
+* Added the file type to the error message when it is not supported.
 
 * Added the `DEMISTO_SDK_CONTENT_PATH` environment variable, implicitly used in various commands.